V0.0.4
First functioning product is complete. It still is missing all the
features, but the kotlin code is done. We can now start working on the
design using html, css, and javascript. A lot has changed from the last
commit. Now instead of searching each file, and reading it's data, the
files are handled by the webClient. Also I added an icon for the time
being, although, it could use some improvement.

IMPORTANT TO NOTE:
- When using html, css, or javascript, you need to reference objects in
  the following way:

  1. If the object is code, store it in the assets folder, and reference
     it with "/assets/{filename}"
  2. If the object is an img, video, sound, etc... then store it in the
     ressources folder, and reference it with
     "/res/{ressourceType}/{filename}"